Add: ヘッダーとthemeにヘッダーの色が反転しない方法を追加した
