+++ frontend/src/theme/theme.js
@@ -3,12 +3,8 @@ import { createTheme } from '@mui/material/styles';
 // Colors
 import { COLORS } from '../style_constants.js';
 
-const theme = createTheme({
-  palette: {
-    green: {
-      main: COLORS.MAIN,
-    },
-  },
+
+/*
   components: {
     // Name of the component
     MuiPaper: {
@@ -16,9 +12,17 @@ const theme = createTheme({
         // Name of the slot
         root: {
           // Some CSS
-          bgcolor: COLORS.MAIN
+          backgroundColor: COLORS.MAIN,
         },
       },
+    },
+  },
+*/
+
+const theme = createTheme({
+  palette: {
+    green: {
+      main: COLORS.MAIN,
     },
   },
 });
